Refactor logger component application
